getting game screen set up.
the road and speed limit sign are up. updates to the assets for street
light with fancy glowing signal and cop lights. spedometer face added as
well as the limit sign face.  limit sign added to the appropriate logic
for when initializing the street. need to get street lights implemented
as well as the car moving across the road.